clean up changes involving the default alg
