+++ lib/OrdinaryDiffEqVerner/src/algorithms.jl
@@ -112,3 +112,8 @@ TruncatedStacktraces.@truncate_stacktrace Vern9 3
 function Vern9(stage_limiter!, step_limiter! = trivial_limiter!; lazy = true)
     Vern9(stage_limiter!, step_limiter!, False(), lazy)
 end
+
+AutoVern6(alg; lazy = true, kwargs...) = AutoAlgSwitch(Vern6(lazy = lazy), alg; kwargs...)
+AutoVern7(alg; lazy = true, kwargs...) = AutoAlgSwitch(Vern7(lazy = lazy), alg; kwargs...)
+AutoVern8(alg; lazy = true, kwargs...) = AutoAlgSwitch(Vern8(lazy = lazy), alg; kwargs...)
+AutoVern9(alg; lazy = true, kwargs...) = AutoAlgSwitch(Vern9(lazy = lazy), alg; kwargs...)
